adding uses to both components

+++ src/App.js
@@ -11,16 +11,37 @@ class App extends Component {
   constructor(props){
     super(props)
     this.state = {
-      theTasks: [],
-      showing: false
+      theTasks: null,
+      showing: false,
+      loggedInUser: null,
     }
   }
 
+
+  fetchUser(){
+    if( this.state.loggedInUser === null ){  
+        axios.get(`http://localhost:5000/api/loggedin`, {withCredentials: true})
+        .then((response)=>{
+            this.setState({
+                theTasks: this.state.theTasks,
+                showing: this.state.showing,
+                loggedInUser:  response.data,
+           }) 
+        })
+        .catch((err)=>{
+            this.setState({
+              theTasks: this.state.theTasks,
+              showing: this.state.showing,
+                loggedInUser:  false,
+           }) 
+        })
+    }
+}
+
   getAllTheTasks(){
-    axios.get("http://localhost:5000/api/tasks")
+    axios.get("http://localhost:5000/api/tasks", {withCredentials: true})
     .then((allTheTasks)=>{
-      // console.log(allTheTasks)
-      this.setState({theTasks: allTheTasks.data, showing: false})
+      this.setState({theTasks: allTheTasks.data, showing: false, loggedInUser: this.state.loggedInUser})
     })
     .catch((err)=>{
       console.log(err)
@@ -45,7 +66,7 @@ class App extends Component {
   }
 
     deleteTask(theIdOfTheTask){
-      axios.post(`http://localhost:5000/api/tasks/delete/${theIdOfTheTask}`, {})
+      axios.post(`http://localhost:5000/api/tasks/delete/${theIdOfTheTask}`, {}, {withCredentials: true})
       .then((response)=>{
         console.log(response);
         this.getAllTheTasks();
@@ -53,52 +74,61 @@ class App extends Component {
       .catch((err)=>{
         console.log(err)
       })
-
-
     }
 
-
-
-
-  showTasks(){
-    if(this.state.theTasks.length === 0){
-      this.getAllTheTasks();
-    }
-    // want to find a better way to address this ^
-    
-
-    return (
-      this.state.theTasks.map((task, index) => {
-        return(
-      <div key={index}>
-        <button onClick={()=>{this.deleteTask(task._id)}} style={{float:'right', backgroundColor: 'red', padding: '10px', margin: '0 5px'}}>
+    seeIfTaskBelongsToUser(task, index){
+      if(this.state.loggedInUser && task.owner == this.state.loggedInUser._id){
+        return (
+          <div>
+          <button onClick={()=>{this.deleteTask(task._id)}} style={{float:'right', backgroundColor: 'red', padding: '10px', margin: '0 5px'}}>
           Delete Task
           </button>
         <button onClick={()=>this.toggleEditForm(index)} style={{float:'right', backgroundColor: 'greenyellow', padding: '10px',  margin: '0 5px'}}> 
         Edit This Task 
         </button>
-      <h3>{task.title}</h3>
-      <p style={{maxWidth: '400px'}} >{task.description} </p>
-      {this.renderForm(index, task._id, task.title, task.description)}
+        </div>
+        )
+      } 
+    }
 
-    
-      </div>
-        ) 
-      })
-    )
+
+
+  showTasks(){
+    if(this.state.theTasks === null){
+      this.getAllTheTasks();
+    }
+
+    if(this.state.theTasks){
+
+      return (
+        this.state.theTasks.map((task, index) => {
+          return(
+        <div key={index}>
+          {this.seeIfTaskBelongsToUser(task, index)}
+        <h3>{task.title}</h3>
+        <p style={{maxWidth: '400px'}} >{task.description} </p>
+        {this.renderForm(index, task._id, task.title, task.description)}
+
+      
+        </div>
+          ) 
+        })
+      )
+    } // closes the if statement
   }
 
 
   render() {
     return (
-    <div className="App">
+      <div className="App">
+      {this.fetchUser()}
     <h1 style={{margin: '80px'}}> The Single Greatest To-Do List In The History of Human History</h1>
     
     <div className="add">
     <AddTask blah={()=>this.getAllTheTasks()}></AddTask>
 
     <User></User>
-    
+
     </div>
 
       <div className="list">
